Make it Trackable (Solution proposal) PT.2
- Remove none-essential comments
- Update File name
- Add more robust hitTest point code (supports over/down)
